requirements.txt
annotated-types==0.6.0
apiview==1.3.25
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-cors-headers==4.3.0
django-import-export==3.3.1
djangorestframework==3.14.0
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
et-xmlfile==1.1.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupPy==1.14
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
preshed==3.0.9
psycopg2==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
selenium==4.15.1
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
tablib==3.5.0
thinc==8.2.1
tqdm==4.66.1
trio==0.23.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.4
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
torch==2.2.0
transformers==4.37.2
scikit-learn==1.3.2
threadpoolctl==3.2.0
